Separate adapter for another file
